Correcion errores ruteo/ se agregar toys-reemplaza balanceados
