--- PupShops/front/src/app/Categorias/Balanceados/Gato/page.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import BalanceadosGatos from "@/components/Productos/balanceados/foodCat"; // Asegúrate de que la ruta sea correcta
-
-const CatFood = () => {
-  return (
-    <div>
-
-      <BalanceadosGatos />
-    </div>
-  );
-};
-
-export default CatFood;
