+++ PupShops/front/src/app/Categorias/Toys/Gato/page.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import ToysCat from "@/components/Productos/Toys/toysCat"
+
+
+
+const juguetesCat = () => {
+  return (
+    <div>
+      <h1>Alimentos para Gatos</h1>
+      <ToysCat />
+    </div>
+  );
+};
+
+export default juguetesCat;
