+++ PupShops/front/src/app/Categorias/Toys/Perro/page.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import ToysDog from "@/components/Productos/Toys/toysDog";
+
+const juguetesCat = () => {
+  return (
+    <div>
+      <h1>Alimentos para Gatos</h1>
+      <ToysDog />
+    </div>
+  );
+};
+
+export default juguetesCat;
