+++ PupShops/front/src/components/Productos/Toys/toysCat.tsx
@@ -1,18 +1,83 @@
-const ToysCat: React.FC = () => {
+"use client";
+
+import { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
+import Image from "next/image";
+import { Product } from "@/Interfaces/Iproducts";
+export default function ToysCat() {
+  const router = useRouter();
+  const [productsCatFood, setProductsCatFood] = useState<Product[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    const fetchProducts = async () => {
+      try {
+        const response = await fetch(
+          "http://localhost:3000/products/child/dcad67a5-b9c6-4f87-8a99-5cf455fdde6f"
+        );
+        if (!response.ok) {
+          throw new Error("Error al cargar los productos");
+        }
+        const data = await response.json();
+        setProductsCatFood(data);
+      } catch (err) {
+        //* Verificación de tipo
+
+        if (err instanceof Error) {
+          setError(err.message);
+        } else {
+          setError("Ocurrió un error desconocido");
+        }
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchProducts();
+  }, []);
+
+  if (loading) {
+    return <div>Cargando...</div>;
+  }
+
+  if (error) {
+    return <div>Error: {error}</div>;
+  }
+
   return (
-    <div>
-      <h2 className="text-lg font-semibold mb-2">Juguetes para Gatos</h2>
-      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
-        {/* Aquí agregarás las tarjetas de juguetes para gatos */}
-        <div className="p-4 border rounded-lg shadow-md">
-          <h3 className="text-md font-bold">Juguete Gato 1</h3>
-          <p className="text-gray-600">Descripción del juguete.</p>
-          <span className="text-teal-600">$Precio</span>
-        </div>
-        {/* Repetir para otros juguetes */}
+    <div className="container mx-auto p-4">
+      <h1 className="text-2xl font-bold mb-4">Balanceados para Gatos</h1>
+      <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4">
+        {productsCatFood.map((product: Product) => (
+          <div
+            key={product.id}
+            className="border p-4 rounded-lg shadow-md cursor-pointer"
+            onClick={() => router.push("/Categorias/Balanceados/Gato")}
+          >
+            <Image
+              src={product.image}
+              alt={product.name}
+              width={300}
+              height={300}
+              className="object-cover rounded-md"
+            />
+            <h2 className="text-lg font-semibold mt-2">{product.name}</h2>
+            <p className="text-gray-700">{product.description}</p>
+            <p className="text-green-600 font-bold">${product.price}</p>
+            <p className="text-gray-500">Stock: {product.stock}</p>
+            <button
+              className="mt-2 w-full bg-blue-500 text-white py-2 rounded-md hover:bg-blue-600 transition"
+              onClick={(e) => {
+                e.stopPropagation(); // Evita que se ejecute el onClick del div
+                console.log(`Agregando ${product.name} al carrito`);
+              }}
+            >
+              Agregar al Carrito
+            </button>
+          </div>
+        ))}
       </div>
     </div>
   );
-};
-
-export default ToysCat;
+}
